EY-1916 Gjor inputfelter for tall til text inputmode number og annen rydding av kode

+++ apps/etterlatte-trygdetid/src/main/resources/db/migration/V2__legg_til_trygdetid_i_trygdetidgrunnlag.sql
@@ -1,1 +1,1 @@
-ALTER TABLE trygdetid_grunnlag ADD COLUMN trygdetid BIGINT;
+ALTER TABLE trygdetid_grunnlag ADD COLUMN trygdetid INTEGER;
